+++ .vscode/spellright.dict
@@ -1,4 +1,5 @@
 abis
 Fastconfirmation
+postconfirm
 testnet
 unstaking
