+++ network/mcr/README.md
@@ -1,0 +1,10 @@
+# MCR Network
+
+This is the MCR Network component.
+
+It is for ???
+
+## Content
+
+- [`components`](./components/README.md): ???
+- [`cli`](./cli/README.md): The CLI for the MCR Network.
