+++ network/mcr/components/eth/anvil/simulations/README.md
@@ -1,0 +1,5 @@
+# Simulations
+
+Simulations are used to simulate the protocol.
+
+- [`mySimulation`](./mySimulation/README.md): Here should go the first simulation.
