+++ network/mcr/components/eth/anvil/tests/README.md
@@ -1,0 +1,6 @@
+## Tests
+
+Tests are used to test the protocol.
+
+- [`basic`](./basic/README.md): Bring up a basic network.
+- [`test2`](./test2/README.md): Stake MOVE tokens to a validator and then post confirm a commitment on an network.
