+++ network/mcr/components/eth/anvil/tests/test2/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "test-test2-mcr-network-anvil-component-core"
+version = { workspace = true }
+edition  = { workspace = true }
+license  = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+publish = { workspace = true }
+rust-version = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+mcr-protocol-client-core-util = { workspace = true }
+
+anyhow = { workspace = true }
+async-stream = { workspace = true }
+async-trait = { workspace = true }
+futures = { workspace = true }
+tokio = { workspace = true }
+tokio-stream = { workspace = true }
+serde_json = { workspace = true }
+mcr-types = { workspace = true }
+mcr-config = { workspace = true }
+kestrel = { workspace = true }
+mcr-protocol-client-eth-core = { workspace = true }
+mcr-protocol-deployer-eth-core = { workspace = true }
+mcr-network-anvil-component-core = { workspace = true }
+strip-ansi-escapes = { workspace = true }
+network-anvil-component-core = { workspace = true }
+secure-signer = { workspace = true }
+secure-signer-loader = { workspace = true }
+
+[dev-dependencies]
+mcr-protocol-client-core-mock = { workspace = true }
+
+[lints]
+workspace = true
